% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drks_de_dates.R
\name{drks_de_dates}
\alias{drks_de_dates}
\title{Download a vector of dates on which a drks.de registry entry was
updated}
\usage{
drks_de_dates(drksid)
}
\arguments{
\item{drksid}{A character string including a well-formed DRKS id,
e.g. "DRKS00005219". (A capitalized "DRKS" followed by eight
numerals with no spaces or hyphens.)}
}
\value{
A vector of ISO-8601 formatted dates corresponding to the
dates on which there were clinical trial history version
updates.
}
\description{
Download a vector of dates on which a drks.de registry entry was
updated
}
\examples{

\dontrun{
versions <- drks_de_dates("DRKS00005219")
}

}
