% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drks_de_version_data.R
\name{drks_de_version_data}
\alias{drks_de_version_data}
\title{Download a registry entry version from DRKS.de}
\usage{
drks_de_version_data(drksid, versionno)
}
\arguments{
\item{drksid}{A character string including a well-formed DRKS id,
e.g. "DRKS00005219". (A capitalized "DRKS" followed by eight
numerals with no spaces or hyphens.)}

\item{versionno}{An integer version number, e.g. 1}
}
\value{
A list containing the overall status, enrolment, start
date, primary completion date, primary completion date type,
minimum age, maximum age, sex, gender-based, accepts healthy
volunteers, inclusion/exclusion criteria, outcome measures,
contacts and sponsors
}
\description{
Download a registry entry version from DRKS.de
}
\examples{
version <- drks_de_version_data ("DRKS00005219", 1)

}
