% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_publications.R
\name{extract_publications}
\alias{extract_publications}
\title{Converts a data frame of the type provided by
\code{clinicaltrials_gov_download()} and converts it to a new data frame
containing the}
\usage{
extract_publications(df, types = c("RESULT", "BACKGROUND", "DERIVED"))
}
\arguments{
\item{df}{A data frame containing the following columns: \code{nctid},
\code{version_number}, \code{total_versions}, \code{version_date},
\code{references}, of the type provided by
\code{clinicaltrials_gov_download()}.}

\item{types}{A list of types to be returned or a character string
if only one type specified, e.g. "RESULT" or c("RESULT",
"BACKGROUND"). Allowed types: "RESULT", "BACKGROUND",
"DERIVED".}
}
\value{
A data frame with all the original columns, as well as an
additional three columns: \code{pmid}, \code{type} and \code{citation}. The
new data frame will have one row per publication.
}
\description{
Converts a data frame of the type provided by
\code{clinicaltrials_gov_download()} and converts it to a new data frame
containing the
}
